Create FeaturedComponent
